+++ naas_drivers/tools/notion.py
@@ -108,11 +108,25 @@ class Notion(InDriver, OutDriver):
         def query(self, database_id, query={}):
             database_id = ensure_database_id(database_id)
             ret = []
-            results = self.client.databases.query(database_id=database_id, **query).get(
-                "results"
-            )
-            for r in results:
-                ret.append(from_dict(data_class=Page, data=r))
+            has_more = True
+            next_cursor = None
+
+            while has_more:
+                if has_more is True and next_cursor is not None:
+                    query["start_cursor"] = next_cursor
+
+                response = self.client.databases.query(database_id=database_id, **query)
+                results = response.get("results")
+
+                for r in results:
+                    ret.append(from_dict(data_class=Page, data=r))
+
+                if response.get("has_more") is True:
+                    has_more = True
+                    next_cursor = response.get("next_cursor")
+                else:
+                    has_more = False
+
             return ret
 
         def create(self, db):
@@ -669,7 +683,7 @@ class Database(__BaseDataClass):
     title: List[RichText]
     properties: object
     parent: Parent
-    #icon: Optional[Emoji] = None  # TODO: Handle File
+    # icon: Optional[Emoji] = None  # TODO: Handle File
     cover: Optional[File] = None
     created_time: str = None
     last_edited_time: str = None
@@ -719,9 +733,9 @@ class Database(__BaseDataClass):
     def from_dict(cls, data):
         return from_dict(data_class=cls, data=data)
 
-    #def set_emoji_icon(
+    # def set_emoji_icon(
     #    self, data
-    #):  # TODO: Fix, seems like there is an issue with notion-client / httpx maybe.
+    # ):  # TODO: Fix, seems like there is an issue with notion-client / httpx maybe.
     #    self.icon = Emoji(data)
 
     def duplicate(self):
@@ -738,7 +752,7 @@ class Database(__BaseDataClass):
         self.properties[col_name] = DatabasePropertyFactory.new(data)
 
     def df(self):
-        pages = self.query()
+        pages = self.query({"page_size": 10})
         frames = []
 
         for page in pages:
@@ -1081,7 +1095,7 @@ class Page(__BaseDataClass):
     parent: Parent
     blocks: Optional[List["Block"]] = field(default_factory=list)
     archived: bool = False
-    #icon: Optional[Emoji] = None  # TODO: Handle File
+    # icon: Optional[Emoji] = None  # TODO: Handle File
     cover: Optional[File] = None
     created_time: str = None
     last_edited_time: str = None